Support to send parameters in the gradle execution.
