+++ gradle.properties
@@ -1,10 +1,10 @@
 #Settings to PivotalTracker - UI
-authentication.baseURL=https://www.pivotaltracker.com
+authenticationBaseURL=https://www.pivotaltracker.com
 user=
 password=
 
 #Settings to PivotalTracker - API
-authentication.baseURI=https://www.pivotaltracker.com/services/v5
+authenticationBaseURI=https://www.pivotaltracker.com/services/v5
 token=
 
 #Selenium settings
