--- src/main/java/org/fundacionjala/pivotal/SuiteListener.java
@@ -1,19 +1,0 @@
-package org.fundacionjala.pivotal;
-
-import org.testng.ISuite;
-import org.testng.ISuiteListener;
-
-/**
- * Created by bruno on 4/9/2017.
- */
-public class SuiteListener implements ISuiteListener {
-    @Override
-    public void onStart(ISuite suite) {
-
-    }
-
-    @Override
-    public void onFinish(ISuite suite) {
-
-    }
-}
