+++ src/main/java/org/fundacionjala/pivotal/testng/SuiteListener.java
@@ -1,0 +1,18 @@
+package org.fundacionjala.pivotal.testng;
+
+import org.testng.ISuite;
+import org.testng.ISuiteListener;
+
+/**
+ * Created by bruno on 4/9/2017.
+ */
+public class SuiteListener implements ISuiteListener {
+    @Override
+    public void onStart(ISuite suite) {
+
+    }
+
+    @Override
+    public void onFinish(ISuite suite) {
+    }
+}
