+++ src/main/java/org/fundacionjala/pivotal/ui/browser/Chrome.java
@@ -8,7 +8,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
  */
 public class Chrome implements Driver {
 
-    private static final String WEB_DRIVER_PATH = "driver/chromedriver.exe";
+    private static final String WEB_DRIVER_PATH = "drivers/chromedriver.exe";
 
     private static final String WEB_DRIVER_KEY = "webdriver.chrome.driver";
 
@@ -16,7 +16,7 @@ public class Chrome implements Driver {
      * {@inheritDoc}
      */
     @Override
-    public final WebDriver initDriver() {
+    public WebDriver initDriver() {
         System.setProperty(WEB_DRIVER_KEY, WEB_DRIVER_PATH);
         return new ChromeDriver();
     }
