--- src/test/java/org/fundacionjala/pivotal/ui/pages/TestSuit.java
@@ -1,45 +1,0 @@
-package org.fundacionjala.pivotal.ui.pages;
-
-import org.testng.ITestContext;
-import org.testng.ITestListener;
-import org.testng.ITestResult;
-
-/**
- * Created by bruno on 4/9/2017.
- */
-public class TestSuit implements ITestListener  {
-    @Override
-    public void onTestStart(ITestResult result) {
-
-    }
-
-    @Override
-    public void onTestSuccess(ITestResult result) {
-
-    }
-
-    @Override
-    public void onTestFailure(ITestResult result) {
-
-    }
-
-    @Override
-    public void onTestSkipped(ITestResult result) {
-
-    }
-
-    @Override
-    public void onTestFailedButWithinSuccessPercentage(ITestResult result) {
-
-    }
-
-    @Override
-    public void onStart(ITestContext context) {
-
-    }
-
-    @Override
-    public void onFinish(ITestContext context) {
-
-    }
-}
