+++ src/test/resources/suite1.xml
@@ -2,10 +2,9 @@
 
 <suite name="Suite1" verbose="1" >
 
-
     <test name="Regression1">
         <classes>
-            <class name="org.fundacionjala.pivotal.ui.pages.project.ProjectManagementTest"/>
+            <class name="org.fundacionjala.pivotal.ui.pages.ProjectManagementTest"/>
             <class name="org.fundacionjala.pivotal.ui.pages.StoryDashBoardTest"/>
         </classes>
     </test>
